refactor(contracts): extract non-visual concerns to src/lib
Step 4.1 of desktop/mobile UI separation refactor.

Extract types, validation helpers, and pricing wizard step graph from
visual portal-family files so the eventual desktop/mobile splits in
Step 4.2 share one source of truth and cannot drift on business logic.

- src/lib/consultation/contracts.ts: ConsultationSelectedServices and
  related payload types (was in booking-modal.tsx)
- src/lib/consultation/booking-helpers.ts: validators / normalizers /
  server-error mapping (was in booking-modal.tsx)
- src/lib/pricing/contracts.ts: FormAnswers, QuestionDef, SelectOption
  (was in pricing-form-modal.tsx)
- src/lib/pricing/wizard.ts: buildAllSteps + option tables — pure layer
  only. JSX-bearing helperText replaced by helperKey strings; the
  visual layer maps keys to React nodes via a local helperContent map.

Updated all consumers in lockstep: pricing-page-client.tsx,
pricing-store.ts, floating-bubble.tsx, booking-modal.tsx,
pricing-form-modal.tsx, plus the legacy-shell copies.

src/lib/ has zero imports from src/components/ (except type-only
imports of data models in pricing-store.ts, which is acceptable).

Plan: ~/.claude/plans/name-codex-review-merry-thunder.md
